Trying to asynchronously: 1. Clear the staging area, notes, and output of Lilypond 2. Run the blocks from the runLogoCommands function, which currently might not be silent, but first trying to implement this, as I read that function but could not find a way to make it silent. Excluding the GUI commands might do it but I don't have much clue on how to do that. Read the code and tried Googling too 3. Save the file
But the problem as of now is that the #2 does not end but the 3 starts, which does not add the content in the saved file.

To eliminate this problem I tried to use callback functions so that it starts only once the function call ends, but it is in itself calling more synchornous functions so the callback comes before it actually ends.

To solve this problem I tried looking at the Timeout function which is supposed to wait in given milliseconds time for the return, but I could not implement or understand that properly