Trying to use setTimeout to wait for the note playing
This is ineffective and not the required solution, but it works. Which indicates that the notes are getting added when they are actually played, which is when the 'notes to play' is printed in the console. This line is on L1951 in logo.js but over there I cannot see when/how is it adding that note to the lilypondStaging